dashboard/requirements.txt
absl-py==0.9.0
bokeh==2.0.0
tornado==5.1.1
pandas==1.0.3
pandas-gbq==0.13.1
google-api-python-client==1.8.0
pymemcache==3.0.0
scipy==1.4.1
PyYAML==5.3.1
python-dateutil==2.8.1
Jinja2>=2.7
numpy==1.18.2
pillow==7.0.0
packaging==20.3
typing_extensions>=3.7.4.1
redis==3.4.1
